requirements.txt
# Core dependencies
Django==5.1.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1

# JWT / Auth
PyJWT==2.9.0

# Utilities
pillow==11.1.0              # For image field support
sqlparse==0.5.3             # Django dependency

# Dev Tools
virtualenv==20.29.2         # To create virtual environments
pipenv==2024.4.1            # Optional: if you're using pipenv
setuptools==75.8.0
filelock==3.17.0
packaging==24.2
platformdirs==4.3.6
certifi==2025.1.31
distlib==0.3.9
asgiref==3.8.1              # Async support required by Django
